Add type check in the `PortNamespace.serialize` method
The serialization is called before the input validation, because the
type of the input might not be correct before serialization. This means
however that if a wrong value is passed for an input port, for example
the user by accident passes a `orm.Dict` instance to `metadata.options`
for a `CalcJob` instead of a plain dict, cryptic messages can be thrown
because the `serialize` method will assume the passed value is iterable.
To prevent this we add an explicit type check before serializing the
port value.
