+++ aiida/backends/tests/engine/test_calc_job.py
@@ -11,6 +11,8 @@
 from __future__ import division
 from __future__ import print_function
 from __future__ import absolute_import
+
+from aiida import orm
 from aiida.backends.testbase import AiidaTestCase
 from aiida.common import exceptions
 from aiida.engine import launch, CalcJob, Process
@@ -18,6 +20,11 @@ from aiida.engine import launch, CalcJob, Process
 
 class TestCalcJob(AiidaTestCase):
     """Test for the `CalcJob` process sub class."""
+
+    @classmethod
+    def setUpClass(cls, *args, **kwargs):
+        super(TestCalcJob, cls).setUpClass(*args, **kwargs)
+        cls.code = orm.Code(remote_computer_exec=(cls.computer, '/bin/true')).store()
 
     def setUp(self):
         super(TestCalcJob, self).setUp()
@@ -59,3 +66,20 @@ class TestCalcJob(AiidaTestCase):
 
         with self.assertRaises(AssertionError):
             launch.run(IncompleteDefineCalcJob)
+
+    def test_invalid_options_type(self):
+        """Verify that passing an invalid type to `metadata.options` raises a `TypeError`."""
+
+        class SimpleCalcJob(CalcJob):
+            """Simple `CalcJob` implementation"""
+
+            @classmethod
+            def define(cls, spec):
+                super(SimpleCalcJob, cls).define(spec)
+
+            def prepare_for_submission(self, folder):
+                pass
+
+        # The `metadata.options` input expects a plain dict and not a node `Dict`
+        with self.assertRaises(TypeError):
+            launch.run(SimpleCalcJob, code=self.code, metadata={'options': orm.Dict(dict={'a': 1})})
